+++ src/debugging/coreclr/ChildProcessProvider.ts
@@ -15,8 +15,9 @@ export interface ProcessProvider {
     exec(command: string, options: ProcessProviderExecOptions): Promise<{ stdout: string, stderr: string }>;
 }
 
+// note: Don't use this as this will be removed/refacotred.
+// use execAsync instead.
 export class ChildProcessProvider implements ProcessProvider {
-
     public get env(): { [key: string]: string | undefined } {
         return process.env;
     }
